fix: tava bugado
