+++ app/src/main/java/com/example/reserveroom/MainActivity.kt
@@ -1,30 +1,15 @@
-import com.google.firebase.database.Exclude
-import com.google.firebase.firestore.IgnoreExtraProperties
-import java.util.HashMap
+package com.example.reserveroom
 
-@IgnoreExtraProperties
-data class Schedules(
-    var time: String = "",
-    var cpf: String = "",
-    var name: String = "",
-    var email: String = "",
-    var available: Boolean = true
-) {
-    // Id é uma propriedade mutável para que possa ser alterada pelo Firebase
-    var id: String? = null
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.google.firebase.FirebaseApp
 
-    // Construtor vazio necessário para usar o Firebase
-    constructor() : this("", "", "", "", true)
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+        FirebaseApp.initializeApp(this)
 
-    // Método para converter para mapa antes de enviar para o Firebase
-    @Exclude
-    fun toMap(): Map<String, Any?> {
-        val result = HashMap<String, Any?>()
-        result["time"] = time
-        result["cpf"] = cpf
-        result["name"] = name
-        result["email"] = email
-        result["available"] = available
-        return result
+
     }
 }
