+++ app/src/main/res/layout/activity_main.xml
@@ -37,7 +37,7 @@
         android:layout_width="388dp"
         android:layout_height="57dp"
         android:layout_marginTop="16dp"
-        android:text="Salas da <Empresa>"
+        android:text="Salas da Empresa"
         android:textAlignment="center"
         android:textSize="36sp"
         android:textStyle="bold"
